fix: icon block generation
